examenSemana16/backend/requirements.txt
asgiref==3.8.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==5.2.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pillow==11.2.1
requests==2.31.0
sqlparse==0.5.3
urllib3==2.0.3
Werkzeug==2.3.6